refactor studentViewTeamBtn to StudentHomePage
